backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.9.2
pydantic-settings==2.5.2

# AI/ML Dependencies
openai==1.10.0
sentence-transformers==2.3.1
torch==2.2.2

# Vector Database
qdrant-client>=1.16.0

# PostgreSQL
psycopg[binary]==3.1.16
psycopg2-binary==2.9.9

# Utilities
python-dotenv==1.0.0
httpx==0.26.0

# NLP for chunking
spacy==3.8.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3

# Database Migrations
alembic==1.13.1

# Rate Limiting
slowapi==0.1.9
